fixes after review: fetcher files renaming
