--- src/fetchers/sushiswap/index.ts
@@ -1,8 +1,0 @@
-import { AnySwapFetcher } from "../AnySwapFetcher";
-
-const symbolToPairIdObj: { [symbol: string]: string } =
-  require("./sushiswap-symbol-to-pair-id.json");
-
-const subgraphUrl = "https://api.thegraph.com/subgraphs/name/sushiswap/exchange";
-
-export default new AnySwapFetcher("sushiswap", subgraphUrl, symbolToPairIdObj);
