+++ src/fetchers/yf-unofficial/YfUnofficialFetcher.ts
@@ -33,5 +33,3 @@ export class YfUnofficialFetcher extends BaseFetcher {
   }
 
 };
-
-export default new YfUnofficialFetcher();
